Add script aws-secureboot-blob
That script can be used to create a AWS binary blob that contains a
pre filled variable store. That blob can be used to create a new AMI
that support secure boot on AWS.
The script is mostly following the commands from [0] but it uses the
files from the Ubuntu shim-signed and secureboot-db packages.

[0]
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/uefi-secure-boot-optionB.html