Add dbxupdate-split.py script
The dbxupdate.bin file (eg. from /usr/share/secureboot/updates/dbx/ )
does contain a signature and the binary data that can be used together
with python-uefivars[0].
The dbxupdate-split.py script can split the file into the signature
and the binary data.
This is similar to what is documented under [1] step 7):

> The dbx change request that you downloaded in the previous step
> is already signed with the Microsoft KEK, so you need to
> strip or unpack it.

[0] https://github.com/awslabs/python-uefivars
[1]
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/aws-binary-blob-creation.html